admn refactor response type

+++ server/src/controllers/admin/admin/delete-admin.ts
@@ -1,11 +1,14 @@
 import { Request } from "express";
 import * as _ from "lodash";
 import { IRemoveAdminById } from "../../../use-cases/admin/delete-admin-by-id";
+import { IGetAdminById } from "../../../use-cases/admin/get-admin-by-id";
 
 export default function makeDeleteAdminController({
   deleteAdminById,
+  getAdminById,
 }: {
   deleteAdminById: IRemoveAdminById;
+  getAdminById: IGetAdminById;
 }) {
   return async function deleteAdminController(
     httpRequest: Request & { context: { validated: { _id: string } } }
@@ -18,6 +21,18 @@ export default function makeDeleteAdminController({
         httpRequest,
         "context.validated"
       );
+
+      const exists = await getAdminById({ id: admin_id });
+      if (!exists) {
+        return {
+          headers,
+          statusCode: 200,
+          body: {
+            is_error: true,
+            message: "Admin does not exist.",
+          },
+        };
+      }
 
       const admin = await deleteAdminById({ id: admin_id });
