+++ server/src/controllers/admin/feedback/delete-feedback.ts
@@ -1,11 +1,14 @@
 import { Request } from "express";
 import * as _ from "lodash";
 import { IRemoveFeedback } from "../../../use-cases/feedback/delete-feedback";
+import { IGetFeedbackById } from "../../../use-cases/feedback/get-feedback-by-id";
 
 export default function makeDeleteFeedbackController({
   deleteFeedbackById,
+  getFeedbackById,
 }: {
   deleteFeedbackById: IRemoveFeedback;
+  getFeedbackById: IGetFeedbackById;
 }) {
   return async function deleteFeedbackController(
     httpRequest: Request & { context: { validated: { _id: string } } }
@@ -18,6 +21,18 @@ export default function makeDeleteFeedbackController({
         httpRequest,
         "context.validated"
       );
+
+      const exists = await getFeedbackById({ id: feedback_id });
+      if (!exists) {
+        return {
+          headers,
+          statusCode: 200,
+          body: {
+            is_error: true,
+            message: "Feedback does not exist.",
+          },
+        };
+      }
 
       const feedback = await deleteFeedbackById({ id: feedback_id });
