+++ server/src/controllers/admin/series/delete-series.ts
@@ -1,11 +1,14 @@
 import { Request } from "express";
 import * as _ from "lodash";
 import { IRemoveSeriesById } from "../../../use-cases/series/delete-series-by-id";
+import { IGetSeriesById } from "../../../use-cases/series/get-series-by-id";
 
 export default function makeDeleteSeriesController({
   deleteSeriesById,
+  getSeriesById,
 }: {
   deleteSeriesById: IRemoveSeriesById;
+  getSeriesById: IGetSeriesById;
 }) {
   return async function deleteSeriesController(
     httpRequest: Request & { context: { validated: { _id: string } } }
@@ -18,6 +21,19 @@ export default function makeDeleteSeriesController({
         httpRequest,
         "context.validated"
       );
+
+      const exists = await getSeriesById({ id: series_id });
+
+      if (!exists) {
+        return {
+          headers,
+          statusCode: 200,
+          body: {
+            is_error: true,
+            message: `Series does not exists.`,
+          },
+        };
+      }
 
       const series = await deleteSeriesById({ id: series_id });
